refactor: extract ParticipantDropdown functions into hook
Move functions, utilities, and types specific to `ParticipantDropdown`s
into own hook, which can be used more portably rather than strictly in
the Take Now modal.
