+++ src/modules/Dashboard/components/TakeNowModal/ParticipantDropdown/index.ts
@@ -1,2 +1,3 @@
 export * from './ParticipantDropdown';
 export * from './ParticipantDropdown.types';
+export * from './ParticipantDropdown.hooks';
